Added octave compatibility to run demo #1
- Fixed save function to save as v7 format if octave
  is used.
- When Octave is used don't use RandStream to set the
  seed.
- When Octave is used don't create a matlabpool, this
  means that octave users currently can't use the
  multicore functionality. But Matlab users are
  still able to.
- Updated readme about octave compatibility

NOTE: There may still be other compatibility issues
I only made it possible to run the Demo on octave.

This patch does not influence any of the existing
matlab behaviour and checks whether the user uses
Octave. It should be a relatively "safe" patch.

I'm still an Octave/Matlab beginner so careful review
would be appreciated.